# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/devices": {
      "post": {
        "operationId": "DevicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Devices"
        ]
      },
      "get": {
        "operationId": "DevicesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Devices"
        ]
      }
    },
    "/devices/{id}": {
      "get": {
        "operationId": "DevicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Devices"
        ]
      },
      "patch": {
        "operationId": "DevicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Devices"
        ]
      },
      "delete": {
        "operationId": "DevicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Devices"
        ]
      }
    },
...
    "/telemetry/device/{deviceId}": {
      "get": {
        "operationId": "TelemetryController_getByDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Telemetry"
        ]
      }
    },
...
    "/driver-activity/rickshaw/{rickshawId}": {
      "get": {
        "operationId": "DriverActivityController_findByRickshaw",
        "parameters": [
          {
            "name": "rickshawId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverActivity"
        ]
      }
    },
    "/driver-activity/rickshaw/{rickshawId}/summary": {
      "get": {
        "operationId": "DriverActivityController_getSummary",
        "parameters": [
          {
            "name": "rickshawId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "example": "DAILY",
              "type": "string",
              "enum": [
                "DAILY",
                "WEEKLY",
                "MONTHLY",
                "QUARTERLY",
                "YEARLY",
                "CUSTOM"
              ]
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-25",
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "example": 6,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          },
          {
            "name": "quarter",
            "required": false,
            "in": "query",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-22",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-29",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverActivity"
        ]
      }
    },
    "/driver-activity/{id}": {
      "get": {
        "operationId": "DriverActivityController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverActivity"
        ]
      },
      "patch": {
        "operationId": "DriverActivityController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverActivity"
        ]
      },
      "delete": {
        "operationId": "DriverActivityController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverActivity"
        ]
      }
    },
...
    "/drivers/{id}": {
      "get": {
        "operationId": "DriversController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      },
      "patch": {
        "operationId": "DriversController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      },
      "delete": {
        "operationId": "DriversController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      }
    },
...
    "/device-provision": {
      "post": {
        "operationId": "DeviceProvisionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceProvisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "DeviceProvision"
        ]
      },
      "get": {
        "operationId": "DeviceProvisionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DeviceProvision"
        ]
      }
    },
    "/device-provision/{id}": {
      "get": {
        "operationId": "DeviceProvisionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DeviceProvision"
        ]
      }
    },
    "/device-provision/{id}/activate": {
      "post": {
        "operationId": "DeviceProvisionController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "DeviceProvision"
        ]
      }
    },
...
    "/driver-documents/driver/{driverId}": {
      "post": {
        "operationId": "DriverDocumentsController_createForDriver",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      },
      "get": {
        "operationId": "DriverDocumentsController_findByDriver",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      }
    },
...
  },
  "info": {
    "title": "AdOnTheGo API",
    "description": "AdOnTheGo Backend APIs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "player-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Enter: ApiKey <your-api-key>"
      }
    },
    "schemas": {
...
      "CreateDeviceDto": {
        "type": "object",
        "properties": {
          "serialNumber": {
            "type": "string",
            "example": "HW-SN-123456"
          },
          "simNumber": {
            "type": "string",
            "example": "919876543210"
          }
        }
      },
      "UpdateDeviceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDriverActivityDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateDriverDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDeviceProvisionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDriverDocumentDto": {
        "type": "object",
        "properties": {
          "documentType": {
            "type": "string",
            "example": "PUC",
            "enum": [
              "AADHAAR_FRONT",
              "AADHAAR_BACK",
              "PAN_CARD",
              "LICENCE_FRONT",
              "LICENCE_BACK",
              "RC_BOOK",
              "INSURANCE",
              "PUC",
              "PERMIT"
            ]
          },
          "fileUrl": {
            "type": "string",
            "example": "puc.pdf"
          },
          "expiryDate": {
            "type": "string",
            "example": "2027-06-26"
          }
        },
        "required": [
          "documentType",
          "fileUrl"
        ]
      },
...
    }
  }
}